+++ Kanikama/Packages/net.shivaduke28.kanikama/Editor/Baking/RendererWithShadowCastingMode.cs
@@ -1,0 +1,27 @@
+﻿using UnityEngine;
+using UnityEngine.Rendering;
+
+namespace Kanikama.Editor.Baking
+{
+    public sealed class RendererWithShadowCastingMode
+    {
+        readonly ObjectHandle<Renderer> handle;
+        readonly ShadowCastingMode shadowCastingMode;
+
+        public RendererWithShadowCastingMode(Renderer renderer)
+        {
+            handle = new ObjectHandle<Renderer>(renderer);
+            shadowCastingMode = handle.Value.shadowCastingMode;
+        }
+
+        public void SetShadowCastingMode(ShadowCastingMode mode)
+        {
+            handle.Value.shadowCastingMode = mode;
+        }
+
+        public void ClearShadowCastingMode()
+        {
+            handle.Value.shadowCastingMode = shadowCastingMode;
+        }
+    }
+}
