+++ Kanikama/Packages/net.shivaduke28.kanikama/Editor/Baking/RendererWithShadowCastingMode.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 677070f5d69f4c9d86ca0ca7decdad60
+timeCreated: 1687266744
